use decimal instead of double for money values since it is more exact

+++ IM_P2/CleanCodeExamples/Product.cs
@@ -1,3 +1,3 @@
 namespace IM_P2.CleanCodeExamples;
 
-public record Product(string Name, double Price);
+public record Product(string Name, decimal Price);
