+++ IM_P2/CleanCodeExamples/ProductParser.cs
@@ -4,5 +4,5 @@ public static class ProductParser
 {
     public static IEnumerable<Product> ParseRawData(IEnumerable<string> lines) => lines
         .Select(line => line.Split(','))
-        .Select(parts => new Product(Name: parts[0], Price: double.Parse(parts[1])));
+        .Select(parts => new Product(Name: parts[0], Price: decimal.Parse(parts[1])));
 }
